Refactor groups functionalities out of issues-viewer as a service (#384)
Currently, the issues-viewer component contains the fields and functions
tracking the groups to be shown and hidden.

Let's refactor out these functionalities into a new service to ensure 
better separations of concerns.
