+++ src/app/issues-viewer/issues-viewer.component.html
@@ -1,5 +1,5 @@
 <div>
-  <div class="loading-spinner" *ngIf="(this.viewService.isChangingRepo | async) || this.groups.length === 0; else elseBlock">
+  <div class="loading-spinner" *ngIf="(this.viewService.isChangingRepo | async) || this.groupService.groups.length === 0; else elseBlock">
     <mat-progress-spinner color="primary" mode="indeterminate" diameter="50" strokeWidth="5"> </mat-progress-spinner>
   </div>
 
@@ -8,15 +8,15 @@
 
     <div class="wrapper">
       <app-card-view
-        *ngFor="let group of groups"
+        *ngFor="let group of this.groupService.groups"
         class="issue-table"
         #card
         [ngStyle]="{ display: card.isLoading || card.issueLength > 0 ? 'initial' : 'none' }"
         [group]="group"
         [headers]="this.displayedColumns"
-        (issueLengthChange)="updateHiddenGroups($event, group)"
+        (issueLengthChange)="this.groupService.updateHiddenGroups($event, group)"
       ></app-card-view>
-      <app-hidden-groups [groups]="this.hiddenGroups"></app-hidden-groups>
+      <app-hidden-groups [groups]="this.groupService.hiddenGroups"></app-hidden-groups>
     </div>
   </ng-template>
 </div>
